requirements.txt
streamlit
python-docx
langchain
langchain-community
langchain-openai
chromadb
tiktoken
openai
